refactor: src/tools/ → src/agent/tools/
Tool definitions belong with the agent layer that registers them.
Move src/tools/types.ts into src/agent/tools/, and convert src/agent/tools.ts
into src/agent/tools/index.ts so future legacy/tools/ migrations have
a clear destination.

After this PR src/tools/ is gone. Layout:

  src/agent/tools/
    index.ts       — createAgentTools, configureToolsLayer, applyToolPolicy
    index.test.ts
    types.ts       — AgentToolSettings (was src/tools/types.ts)

src/legacy/tools/ stays for now (exec/web/react/git/github); will move
into src/agent/tools/ in follow-up work.
