+++ src/agent/tools/index.test.ts
@@ -3,9 +3,9 @@ import {
   createTimeTool,
   createAgentTools,
   applyToolPolicy,
-} from "./tools.js";
-import { createWebFetchTool } from "../legacy/tools/web.js";
-import { ProcessRegistry } from "../legacy/tools/exec.js";
+} from "./index.js";
+import { createWebFetchTool } from "../../legacy/tools/web.js";
+import { ProcessRegistry } from "../../legacy/tools/exec.js";
 
 function getText(result: { content: Array<{ type: string; text?: string }> }): string {
   const block = result.content.find((c) => c.type === "text");
