+++ src/app.ts
@@ -8,7 +8,7 @@ import { createLogger } from "./logging/logger.js";
 import { LazyTransportContext } from "./legacy/tools/react.js";
 import { ProcessRegistry } from "./legacy/tools/exec.js";
 import type { AgentTool } from "@mariozechner/pi-agent-core";
-import { configureToolsLayer, shutdownToolsLayer } from "./agent/tools.js";
+import { configureToolsLayer, shutdownToolsLayer } from "./agent/tools/index.js";
 import {
   ensureDirectories,
   resolveAgentWorkspacePath,
